pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "schoolphotoID"
version = "1.0.0"
description = "School Photo Face Recognition and Identification Desktop Application"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "SchoolFacerecognition"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Win32",
    "Environment :: X11",
    "Environment :: MacOS",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Home",
    "Topic :: Utility",
    "Topic :: Multimedia :: Graphics",
]
dependencies = [
    "PyQt6>=6.5.0",
    "face-recognition>=1.3.0",
    "dlib>=19.24.0",
    "numpy>=1.24.0",
    "opencv-python>=4.8.0",
    "Pillow>=10.0.0",
]

[project.scripts]
schoolphotoID = "schoolphotoID.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["schoolphotoID*"]